Add shared folder for client, server in repo - Add ESlint files
