+++ package.json
@@ -4,11 +4,7 @@
   "description": "A simple Tic-Tac-Toe game",
   "main": "index.js",
   "scripts": {
-    "start": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "test": "npx cypress open",
-    "deploy_dist": "git subtree push --prefix dist origin gh-pages"
+    "test": "npx cypress open"
   },
   "repository": {
     "type": "git",
@@ -24,11 +20,8 @@
     "url": "https://github.com/akshayp-mangospring/tic-tac-toe/issues"
   },
   "homepage": "https://github.com/akshayp-mangospring/tic-tac-toe#readme",
-  "dependencies": {
-    "socket.io-client": "^4.5.1",
-    "vite": "^2.9.13"
-  },
   "devDependencies": {
+    "cypress": "^10.3.0",
     "eslint": "^8.19.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.26.0"
