+++ shared/game.js
@@ -1,0 +1,32 @@
+const setupGameState = (rowSize = 3) => {
+  const boardSize = rowSize ** 2;
+  const gs = Array(boardSize).fill(null);
+  const minCellsToFill = rowSize * 2 - 1;
+  let cellsFilledCount = 0;
+
+  return Object.freeze({
+    cells: gs,
+    setCell: (i, v) => {
+      gs[i] = v;
+      cellsFilledCount += 1;
+    },
+    canComputeWinner: cellsFilledCount >= minCellsToFill,
+    isBoardFilled: cellsFilledCount === boardSize,
+  });
+};
+
+const getWinCombos = () => Object.freeze([
+  [0, 1, 2],
+  [3, 4, 5],
+  [6, 7, 8],
+  [0, 3, 6],
+  [1, 4, 7],
+  [2, 5, 8],
+  [0, 4, 8],
+  [2, 4, 6],
+]);
+
+module.exports = {
+  setupGameState,
+  getWinCombos,
+};
